+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
     - Add configurable auto jump to exceptions
     - New eql-query-pprint visualizer
+    - New webpage visualizer
     - Add support for setting the visualizer via flow-storm.api/data-window-push-val
     - Add debugger/bookmark to set bookmarks from code and also quickly jump to the first one
     - Enable multiple ClojureScript runtimes <> multiple debuggers via "flowstorm_ws_port"